Renames almost all files for clarity and pytest
